Remove extra controlflow && update parser option (validation) && remove doctest
