+++ mlbench/refimpls/pytorch/controlflow/__init__.py
@@ -1,7 +1,7 @@
 def get_controlflow(options):
-    if options.libsvm_dataset:
-        from .linear_model_controlflow import Train
-        return Train()
-    else:
-        from .controlflow import TrainValidation
-        return TrainValidation()
+    # if options.libsvm_dataset:
+    #     from .linear_model_controlflow import Train
+    #     return Train()
+
+    from .controlflow import TrainValidation
+    return TrainValidation()
