+++ mlbench/refimpls/pytorch/models/linear_models.py
@@ -27,7 +27,7 @@ class RidgeRegression(torch.nn.Module):
     """
 
     def __init__(self, n_features):
-        super(LogisticRegression, self).__init__()
+        super(RidgeRegression, self).__init__()
         self.linear = torch.nn.Linear(n_features, 1, bias=False)
 
     def forward(self, x):
