+++ mlbench/refimpls/pytorch/optim/lr.py
@@ -261,5 +261,4 @@ def get_scheduler(options, optimizer):
 
 
 if __name__ == '__main__':
-    import doctest
-    doctest.testmod()
+    pass
