+++ mlbench/refimpls/pytorch/utils/utils.py
@@ -1,8 +1,8 @@
-def convert_dtype(options, obj):
+def convert_dtype(dtype, obj):
     # The object should be a ``module`` or a ``tensor``
-    if options.dtype == 'fp32':
+    if dtype == 'fp32':
         return obj.float()
-    elif options.dtype == 'fp64':
+    elif dtype == 'fp64':
         return obj.double()
     else:
-        raise NotImplementedError
+        raise NotImplementedError('dtype {} not supported.'.format(dtype))
